// src/types/Districts.ts
export interface DistrictResponse {
    state_code: number;
    state_name_english: string;
    state_name_local: string;
    state_census2011_code: string;
    district_code: number;
    district_name_english: string;
    district_name_local: string;
    district_census2011_code: string;
}

export interface DistrictRecord {
    state_code: number;
    state_name_english: string;
    state_name_local: string;
    district_code: number;
    district_name_english: string;
    district_name_local: string;
}